requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
coverage==6.4.2
Django==3.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
flake8==4.0.1
idna==3.3
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.4.0
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.10
